requirements.txt
temporalio
python-dotenv
telethon
slack-sdk
aiohttp
anthropic
